Enable automatic updating of instance profile credentials, work around issue with signer in some regions
This is a simplified approach to fixing Jenkins issue #26854. Trying both signers (default and QueryStringSigner) might be a little heavy-handed and I am not exactly sure why this is necessary since the Java API should be able to figure out the correct signer to be used per region. Be it as it may, it works for me.